조회수

+++ src/controllers/videoController.js
@@ -10,29 +10,39 @@ export const home = async (req, res) => {
 export const watch = async (req, res) => {
   const { id } = req.params;
   const video = await Video.findById(id).populate("owner");
-
-  if (video) {
-    return res.render("watch", { pageTitle: video.title, video });
+  if (!video) {
+    return res.status(400).render("404", { pageTitle: "video not found" });
   }
-  return res.status(400).render("404", { pageTitle: "video not found" });
+  return res.render("watch", { pageTitle: video.title, video });
 };
 
 export const getEdit = async (req, res) => {
   const { id } = req.params;
+  const {
+    user: { _id },
+  } = req.session;
   const video = await Video.findById(id);
   if (!video) {
     return res.render("404", { pageTitle: "video not found" });
   }
-
+  if (video.owner.toString() !== String(_id)) {
+    return res.status(403).redirect("/");
+  }
   return res.render("edit", { pageTitle: `Editing ${video.title}`, video });
 };
 
 export const postEdit = async (req, res) => {
   const { id } = req.params;
+  const {
+    user: { _id },
+  } = req.session;
   const { title, description, hashtags } = req.body;
   const video = await Video.exists({ _id: id });
   if (!video) {
     return res.status(400).render("404", { pageTitle: "video not found" });
+  }
+  if (video.owner.toString() !== String(_id)) {
+    return res.status(403).redirect("/");
   }
   await Video.findByIdAndUpdate(id, {
     title,
@@ -75,7 +85,18 @@ export const postUpload = async (req, res) => {
 
 export const deleteVideo = async (req, res) => {
   const { id } = req.params;
+  const {
+    user: { _id },
+  } = req.session;
+  const video = await Video.findById(id);
+  if (!video) {
+    return res.status(404).render("404", { pageTitle: "Video not found." });
+  }
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect("/");
+  }
   await Video.findByIdAndDelete(id);
+  z;
   return res.redirect("/");
 };
 
@@ -92,4 +113,18 @@ export const search = async (req, res) => {
   }
 
   return res.render("search", { pageTitle: "Search", videos });
+};
+
+export const registerView = async (req, res) => {
+  const { id } = req.params;
+  console.log("registerView");
+  const video = await Video.findById(id);
+  console.log(video);
+  if (!video) {
+    return res.sendStatus(404);
+  }
+
+  video.meta.views = video.meta.views + 1;
+  await video.save();
+  return res.sendStatus(200);
 };
